# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Management Platform API",
    "version": "1.0.0",
    "description": "Multi-tenant school management API. Every protected request runs: request id → rate limit → authentication → tenant context → permission → entitlement → validation → handler. Errors use stable machine-readable codes."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "PERMISSION_DENIED"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "Tokens": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "number"
          },
          "refresh_token": {
            "type": "string"
          },
          "refresh_expires_at": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "refresh_token",
          "refresh_expires_at",
          "session_id"
        ]
      }
    },
...
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Sign in with email and password",
        "description": "Access: **public**",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@school.test"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "membership_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "device": {
                    "type": "object",
                    "properties": {
                      "device_name": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "device_type": {
                        "type": "string",
                        "enum": [
                          "web",
                          "android",
                          "ios",
                          "desktop",
                          "other"
                        ]
                      }
                    }
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Tokens"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "membership": {}
                          }
                        }
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission, entitlement or tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "summary": "End the current session",
        "description": "Access: **authenticated**",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission, entitlement or tenant access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation or business-rule error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}